move iterator tests
